restructure, need tests and smaller functions
