+++ .gitignore
@@ -1,3 +1,13 @@
+.vscode/*
+!.vscode/settings.json
+!.vscode/tasks.json
+!.vscode/launch.json
+!.vscode/extensions.json
+*.code-workspace
+
+# Local History for Visual Studio Code
+.history/
+
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
