--- mysgen/templates/archive.html
@@ -1,51 +1,0 @@
-{% extends "base/base.html" %}
-{% block title %}{{ SITENAME }} - Archive{% endblock %}
-{% block content %}
-{% set pageName = "archive" %}
-{% if MENUITEMS %}
-{% for title, link in MENUITEMS %}
-{% if title == "archive" %}
-    <div id="page_{{title}}" class="page" {% if title != pageName %}style="display: none"{% endif %}>
-      <div id="content_{{title}}" class="content {% if title == pageName %}ajaxHook{% endif %}">
-        <div id="ajax_content_{{title}}" class="ajax_content">
-          <!--<ol id="category-list" class="ordered-list" data-list-length="{{ ALLCATEGORIES|count }}">
-            {% for cat in ALLCATEGORIES %}
-            <li>
-              {{ cat }}
-            </li>
-            {% endfor %}
-          </ol>
-          <ol id="tag-list" class="ordered-list" data-list-length="{{ ALLTAGS|count }}">
-            {% for tag in ALLTAGS %}
-            <li>
-              {{ tag }}
-            </li>
-            {% endfor %}
-          </ol>-->
-          <ol id="post-list" class="ordered-list" data-list-length="{{ articles|count }}">
-            {% for article in articles %}
-            {% include 'base/archive_kernel.html' %}
-            {% endfor %}
-          </ol>
-        </div>
-      </div>
-    </div>
-    {% else %}
-    <div id="page_{{title}}" class="page" {% if title != pageName %}style="display: none"{% endif %}>
-      <div id="content_{{title}}" class="content {% if title == pageName %}ajaxHook{% endif %}">
-        <script>
-          $("#content_{{title}}").load("{{ SITEURL }}{{ link }} #ajax_content_{{title}}");
-        </script>
-      </div>
-    </div>
-{% endif %}
-{% endfor %}
-{% endif %}
-
-    <!-- Animate to page -->
-    <script>
-    // Which page is this
-    whichPageActive = 2;
-    </script>
-    <!-- End Animate to page -->
-{% endblock %}
