--- mysgen/templates/article.html
@@ -1,48 +1,0 @@
-{% extends "base/base.html" %}
-{% block head %}
-{{ super() }}
-
-{% for keyword in articles.keywords %}
-<meta name="keywords" contents="{{keyword}}" />
-{% endfor %}
-
-{% for description in articles.description %}
-<meta name="description" contents="{{description}}" />
-{% endfor %}
-
-{% for tag in articles.meta.tags %}
-<meta name="tags" contents="{{tag}}" />
-{% endfor %}
-
-{% endblock %}
-{% block content %}
-{% set pageName = "archive" %}
-{% if MENUITEMS %}
-{% for title, link in MENUITEMS %}
-{% if title == "archive" %}
-    <div id="page_{{title}}" class="page" {% if title != pageName %}style="display: none"{% endif %}>
-      <div id="content_{{title}}" class="content {% if title == pageName %}ajaxHook{% endif %}">
-        <div id="ajax_content_{{title}}" class="ajax_content">
-          {% include 'base/article_kernel.html' %}
-        </div>
-      </div>
-    </div>
-{% else %}
-    <div id="page_{{title}}" class="page" {% if title != pageName %}style="display: none"{% endif %}>
-      <div id="content_{{title}}" class="content {% if title == "pageName" %}ajaxHook{% endif %}">
-        <script>
-          $("#content_{{title}}").load("{{ SITEURL }}{{ link }} #ajax_content_{{title}}");
-        </script>
-      </div>
-    </div>
-{% endif %}
-{% endfor %}
-{% endif %}
-
-    <!-- Animate to page -->
-    <script>
-    // Which page is this
-    whichPageActive = 2;
-    </script>
-    <!-- End Animate to page -->
-{% endblock %}
