--- mysgen/templates/base/archive_kernel.html
@@ -1,9 +1,0 @@
-<li class="list-style list-style-tiles" data-list="{{ article.date.strftime('%Y-%m-%d') }}" data-month="{{ article.date.strftime('%m') }}">
-  <a id="{{ article.slug }}" href="{{SITEURL}}{{ article.url }}" rel="bookmark" title="Permalink to {{ article.title|striptags }}">
-    <article class="hentry">
-      <header>
-        {{ article.date.strftime('%Y-%m-%d') }}<h3 class="entry-title">{{ article.title }}</h3>
-      </header>
-    </article>
-  </a>
-</li>
