--- mysgen/templates/base/article_kernel.html
@@ -1,38 +1,0 @@
-          <article class="article">
-            {% if articles.meta.image %}
-            <figure>
-              <img class="post_image" src="{{SITEURL}}{{path}}/{{articles.meta.image}}"/>
-            </figure>
-            {% endif %}
-            <header>
-              <h1 class="entry-title">{{ articles.meta.title }}</h1>
-
-              <span class="published" title="{{ articles.meta.date.isoformat() }}">
-                {{ articles.meta.date.strftime("%Y-%m-%d") }}
-              </span>
-              <!-- {% if articles.modified %}
-              <span class="modified" title="{{ articles.modified.isoformat() }}">
-                Updated {{ articles.modified.strftime("%Y-%m-%d") }} |
-              </span>
-              {% endif %}
-              <span class="posted_in">
-                {{ articles.meta.category }} |
-                {% for tag_ in range(0, articles.meta.tags|length) %}
-                {% if tag_ == articles.meta.tags|length - 1 %}
-                  {{ articles.meta.tags[tag_] }}
-                {% elif tag_ == (articles.meta.tags|length - 2) %}
-                  {{ articles.meta.tags[tag_] }} |
-                {% else %}
-                  {{ articles.meta.tags[tag_] }} |
-                {% endif %}
-                {% endfor %}
-              </span> -->
-
-            </header>
-            <div class="article-content hyphenate">
-              <div class="article_body">
-                {{ articles.content }}
-              </div>
-            </div>
-            <footer class="post-info"></footer>
-          </article>
