--- mysgen/templates/page.html
@@ -1,53 +1,0 @@
-{% extends "base/base.html" %}
-{% block content %}
-{% if MENUITEMS %}
-{% for title, link in MENUITEMS %}
-    <div id="page_{{title}}" class="page" {% if title != page_name %}style="display: none"{% endif %}>
-      <div id="content_{{title}}" class="content {% if title == page_name %}ajaxHook{% endif %}">
-        {% if not (title + ".md" in pages.keys()) %}
-        <script>
-           $("#content_{{title}}").load("{{ SITEURL }}{{ link }} #ajax_content_{{title}}");
-        </script>
-        {% else %}
-        <div id="ajax_content_{{ pages[title+".md"].meta.title.lower() }}" class="ajax_content">
-          {% if title == "projects"%}
-          <ol id="post-list" class="ordered-list" data-list-length="{{ articles|count }}">
-            {% for article in articles %}
-            {% if article.category == "Projects" %}
-            {% include 'base/archive_kernel.html' %}
-            {% endif %}
-            {% endfor %}
-          </ol>
-          {% elif title == "personal"%}
-          <ol id="post-list" class="ordered-list" data-list-length="{{ articles|count }}">
-            {% for article in articles %}
-            {% if article.category == "Personal" %}
-            {% include 'base/archive_kernel.html' %}
-            {% endif %}
-            {% endfor %}
-          </ol>
-          {% else %}
-          <h1 style="display: none">{{ pages[title+".md"].meta.title.lower() }}</h1>
-          <div class="entry-content-page hyphenate">
-            {{ pages[title+".md"].content }}
-          </div>
-          {% endif %}
-          <!-- This is ugly but I haven't found any better way of doing things! Can use JavaScript but users witout it will lose content. -->
-        </div>
-        {% endif %}
-      </div>
-    </div>
-{% endfor %}
-{% endif %}
-
-    <!-- Animate to page -->
-    <script>
-    // Which page is this
-    for (var i = 1; i <= pages.length; i++) {
-      if ( "menu_{{page_name}}" == "menu_" + pages[i-1]) {
-        whichPageActive = i;
-      }
-    }
-    </script>
-    <!-- End Animate to page -->
-{% endblock %}
